Lib Types Checking
Brings over the ability to validate the types of the `.d.ts` files in this package with `tsc`.

Unfortunately, it looks like the TypeScript standard library is also being checked-against, so errors for those are popping up in the type checking too. I think it has to do with overlap of the definition files, since they define some of the same or similar variables and interfaces.

One way I may try to counter this is to see if these definitions can instead be written in full `.ts` files, and can be built down to `.d.ts` ones instead, allowing only the `./src`-ish kinds of files in this library to be checked properly.

I want to ensure you can still import individual types from the library too, that's an important part. Say if you wanted to only use the types from `/// <reference types="new-javascript/webaudio/audio-worklet.d.ts"/>`.

https://github.com/BenjaminAster/TypeScript-types-for-new-JavaScript/issues/4